Restructured by removing superadmin andvoters can register by themselves

+++ demo/src/main/java/com/kathir/demo/repository/CandidateRepository.java
@@ -1,8 +1,8 @@
 package com.kathir.demo.repository;
 
-import com.kathir.demo.modules.Candidate;
+import com.kathir.demo.models.Candidate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CandidateRepository extends JpaRepository<Candidate,Long> {
+public interface CandidateRepository extends JpaRepository<Candidate,Integer> {
 
 }
